+++ src/api/consumer/lib/exceptions/diagnosis-not-found.js
@@ -1,0 +1,11 @@
+/** @flow */
+import AbstractError from '../../../../error/abstract-error';
+
+export default class DiagnosisNotFound extends AbstractError {
+  diagnosisName: string;
+
+  constructor(diagnosisName: string) {
+    super();
+    this.diagnosisName = diagnosisName;
+  }
+}